Add LICENSE file
A conversation among two friends sparked an idea. The idea took on different shape as the days went on, as the code and files began to fill up the project folder. I soon realized I was making something that should've been made a long time ago. I offer this to the community freely. This is only a very small beginning. I have very big plans for this project. I think everyone will enjoy what is coming next. A little internet meme culture mixed together with a new type of package manager. One completely written in Zig.